requirements.txt
openai
pyyaml
python-dotenv